+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/account/AccountTag.kt
@@ -4,7 +4,5 @@ package co.electriccoin.zcash.ui.screen.account
  * These are only used for automated testing.
  */
 object AccountTag {
-    const val STATUS_VIEWS = "status_views"
-    const val SINGLE_LINE_TEXT = "single_line_text"
-    const val FIAT_CONVERSION = "fiat_conversion"
+    const val BALANCE_VIEWS = "balance_views"
 }