+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/account/AndroidAccount.kt
@@ -6,15 +6,10 @@ import androidx.activity.ComponentActivity
 import androidx.activity.viewModels
 import androidx.compose.runtime.Composable
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import co.electriccoin.zcash.ui.common.viewmodel.CheckUpdateViewModel
 import co.electriccoin.zcash.ui.common.viewmodel.WalletViewModel
-import co.electriccoin.zcash.ui.configuration.ConfigurationEntries
-import co.electriccoin.zcash.ui.configuration.RemoteConfig
 import co.electriccoin.zcash.ui.design.component.CircularScreenProgressIndicator
 import co.electriccoin.zcash.ui.screen.account.view.Account
 import co.electriccoin.zcash.ui.screen.settings.viewmodel.SettingsViewModel
-import co.electriccoin.zcash.ui.screen.update.AppUpdateCheckerImp
-import co.electriccoin.zcash.ui.screen.update.model.UpdateState
 
 @Composable
 internal fun WrapAccount(
@@ -23,25 +18,12 @@ internal fun WrapAccount(
     goBalances: () -> Unit,
     goSettings: () -> Unit,
 ) {
-    // Show information about the app update, if available
-    val checkUpdateViewModel by activity.viewModels<CheckUpdateViewModel> {
-        CheckUpdateViewModel.CheckUpdateViewModelFactory(
-            activity.application,
-            AppUpdateCheckerImp.new()
-        )
-    }
-    val updateAvailable =
-        checkUpdateViewModel.updateInfo.collectAsStateWithLifecycle().value.let {
-            it?.appUpdateInfo != null && it.state == UpdateState.Prepared
-        }
-
     val walletViewModel by activity.viewModels<WalletViewModel>()
     val walletSnapshot = walletViewModel.walletSnapshot.collectAsStateWithLifecycle().value
 
     val settingsViewModel by activity.viewModels<SettingsViewModel>()
 
     val isKeepScreenOnWhileSyncing = settingsViewModel.isKeepScreenOnWhileSyncing.collectAsStateWithLifecycle().value
-    val isFiatConversionEnabled = ConfigurationEntries.IS_FIAT_CONVERSION_ENABLED.getValue(RemoteConfig.current)
 
     if (null == walletSnapshot) {
         // TODO [#1146]: Consider moving CircularScreenProgressIndicator from Android layer to View layer
@@ -51,9 +33,7 @@ internal fun WrapAccount(
     } else {
         Account(
             walletSnapshot = walletSnapshot,
-            isUpdateAvailable = updateAvailable,
-            isKeepScreenOnDuringSync = isKeepScreenOnWhileSyncing,
-            isFiatConversionEnabled = isFiatConversionEnabled,
+            isKeepScreenOnWhileSyncing = isKeepScreenOnWhileSyncing,
             goBalances = goBalances,
             goHistory = goHistory,
             goSettings = goSettings,