+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/balances/AndroidBalances.kt
@@ -6,19 +6,45 @@ import androidx.activity.ComponentActivity
 import androidx.activity.viewModels
 import androidx.compose.runtime.Composable
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import co.electriccoin.zcash.ui.common.viewmodel.CheckUpdateViewModel
 import co.electriccoin.zcash.ui.common.viewmodel.WalletViewModel
+import co.electriccoin.zcash.ui.configuration.ConfigurationEntries
+import co.electriccoin.zcash.ui.configuration.RemoteConfig
 import co.electriccoin.zcash.ui.screen.balances.view.Balances
+import co.electriccoin.zcash.ui.screen.settings.viewmodel.SettingsViewModel
+import co.electriccoin.zcash.ui.screen.update.AppUpdateCheckerImp
+import co.electriccoin.zcash.ui.screen.update.model.UpdateState
 
 @Composable
 internal fun WrapBalances(
     activity: ComponentActivity,
     goSettings: () -> Unit,
 ) {
+    // Show information about the app update, if available
+    val checkUpdateViewModel by activity.viewModels<CheckUpdateViewModel> {
+        CheckUpdateViewModel.CheckUpdateViewModelFactory(
+            activity.application,
+            AppUpdateCheckerImp.new()
+        )
+    }
+    val isUpdateAvailable =
+        checkUpdateViewModel.updateInfo.collectAsStateWithLifecycle().value.let {
+            it?.appUpdateInfo != null && it.state == UpdateState.Prepared
+        }
+
+    val settingsViewModel by activity.viewModels<SettingsViewModel>()
+
+    val isKeepScreenOnWhileSyncing = settingsViewModel.isKeepScreenOnWhileSyncing.collectAsStateWithLifecycle().value
+    val isFiatConversionEnabled = ConfigurationEntries.IS_FIAT_CONVERSION_ENABLED.getValue(RemoteConfig.current)
+
     val walletViewModel by activity.viewModels<WalletViewModel>()
     val walletSnapshot = walletViewModel.walletSnapshot.collectAsStateWithLifecycle().value
 
     Balances(
-        walletSnapshot = walletSnapshot,
+        isFiatConversionEnabled = isFiatConversionEnabled,
+        isKeepScreenOnWhileSyncing = isKeepScreenOnWhileSyncing,
+        isUpdateAvailable = isUpdateAvailable,
         onSettings = goSettings,
+        walletSnapshot = walletSnapshot,
     )
 }